refactor(providers): reloca o Instagram (background) para src/providers/instagram/
Move processInstagramCapture e helpers de export/comments visíveis do controller
para o módulo do provider, seguindo o template de X/LinkedIn. getEndpointStore e
recordRawPayload passam para src/background/store.ts (compartilhados, evitando ciclo).
O controller passa a importar a faceta e os helpers de export do novo módulo.

Refs #5 #7
